refactor: Add all components to AboutPageComponent
