+++ wealth-sync/src/components/AboutPageComponents/SocialMediaLinksCard.tsx
